WindshaftWorker added
